move test code to P0475R1_P0591R4_uses_allocator_construction
