+++ tests/std/tests/P0475R1_P0591R4_uses_allocator_construction/test.cpp
@@ -179,6 +179,17 @@ void test_GH_2021() { // COMPILE-ONLY
     tags[0];
 }
 
+struct MoveOnlyType {
+    MoveOnlyType()               = default;
+    MoveOnlyType(MoveOnlyType&&) = default;
+};
+
+void test_LWG3527() { // COMPILE-ONLY
+    std::allocator<MoveOnlyType> alloc;
+    [[maybe_unused]] auto p = pair<MoveOnlyType&&, MoveOnlyType&&>{MoveOnlyType{}, MoveOnlyType{}};
+    [[maybe_unused]] auto t = uses_allocator_construction_args<pair<MoveOnlyType&&, MoveOnlyType&&>>(alloc, move(p));
+}
+
 int main() {
     test_P0475R1();
